[KYUUBI #7270]Enable BasicAuth for Prometheus metrics endpoint
